refactor: rename LicensePlate component
