+++ components/LicensePlate/LicensePlate.tsx
