+++ components/LicensePlate/index.ts
@@ -1,0 +1,1 @@
+export { default as LicensePlate } from "./LicensePlate";
